+++ src/store/formSlice.ts
@@ -1,16 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-
-export interface FormData {
-  name: string;
-  age: number | null;
-  country: string;
-  gender: string;
-  email: string;
-  password: string;
-  confirmPassword: string;
-  file: File | string | null;
-  accept: boolean;
-}
+import { FormData } from '../types/formData.ts';
 
 interface FormState {
   uncontrolledFormData: FormData;
